+++ NEWS.md
@@ -1,12 +1,16 @@
+## `sdam` v0.4.0
+
+* New functions `plot.map()` and `dts()`. Datasets `rpm` and `rpmcd`
+
 ## `sdam` v0.3.0
 
-* New functions `prex()` and `plot.dates()`.
+* New functions `prex()` and `plot.dates()`. Dataset `rp`
 
 ## `sdam` v0.2.0
 
-* New function `edhw()` for manipulation of `EDH`.
+* New function `edhw()` for manipulation of `EDH`
 
 ## `sdam` v0.1.0
 
-* First functions `get.edh()`, `get.edhw()`, `request()`, `simil()`, and `EDH` dataset.
+* First functions `get.edh()`, `get.edhw()`, `request()`, `simil()`, and `EDH` dataset